PlaidAPIImpl/requirements.txt
amqp==5.0.6
anyjson==0.3.3
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.0
cffi==1.15.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==35.0.0
Django==3.2.9
django-rest-knox==4.1.0
djangorestframework==3.12.4
kombu==5.2.1
nulltype==2.3.1
plaid-python==8.6.0
prompt-toolkit==3.0.22
pycparser==2.21
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
